pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uspto-enriched-citation-mcp"
version = "1.0.0"
description = "MCP server for USPTO Enriched Citation API v3 with progressive disclosure"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "John Walkoe" }
]

dependencies = [
    "mcp>=1.1.2",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "click>=8.2.1",
    "structlog>=24.1.0",
    "aiohttp>=3.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "detect-secrets>=1.5.0",
    "pre-commit>=3.5.0",
    "mypy>=1.18.2",
    "black>=25.9.0",
    "ruff>=0.14.3",
]

[project.scripts]
uspto-enriched-citation-mcp = "uspto_enriched_citation_mcp.main:main"
